pyproject.toml
[project]
name = "data-transformation-agent"
version = "0.1.0"
description = "Python document and data transformation agent using the official Claude Agent SDK"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "claude-agent-sdk==0.1.63",
  "mcp==1.27.0",
  "fastapi==0.136.0",
  "uvicorn==0.44.0",
  "flask==3.1.3",
  "python-multipart==0.0.26",
  "python-dotenv==1.2.2",
  "pydantic-settings==2.13.1",
  "httpx==0.28.1",
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"
